Style: Card 폴더 생성 및 구조 변경
